componentização ok
