+++ src/components/SideBar.tsx
@@ -1,3 +1,64 @@
-export function SideBar() {
-  // Complete aqui
+import { useEffect, useState } from 'react';
+
+import { Button } from '../components/Button';
+
+import '../styles/global.scss';
+import '../styles/sidebar.scss';
+
+import { api } from '../services/api';
+
+interface GenreResponseProps {
+  id: number;
+  name: 'action' | 'comedy' | 'documentary' | 'drama' | 'horror' | 'family';
+  title: string;
+}
+
+interface MovieProps {
+  imdbID: string;
+  Title: string;
+  Poster: string;
+  Ratings: Array<{
+    Source: string;
+    Value: string;
+  }>;
+  Runtime: string;
+}
+
+interface SideBarProps {
+  handleClickButton: (id: number) => void,
+  selectedGenreId: number,
+  // selectedGenre: number
+}
+
+export function SideBar({handleClickButton, selectedGenreId}: SideBarProps) {
+  const [genres, setGenres] = useState<GenreResponseProps[]>([]);
+
+  useEffect(() => {
+    api.get<GenreResponseProps[]>('genres').then(response => {
+      setGenres(response.data);
+    });
+  }, []);
+
+  return (
+    <div style={{ display: 'flex', flexDirection: 'row' }}>
+      <nav className="sidebar">
+        <span>Watch<p>Me</p></span>
+
+        <div className="buttons-container">
+          {genres.map(genre => (
+            <Button
+              key={String(genre.id)}
+              title={genre.title}
+              iconName={genre.name}
+              onClick={() => handleClickButton(genre.id)}
+              selected={selectedGenreId === genre.id}
+            />
+          ))}
+        </div>
+
+      </nav>
+
+
+    </div>
+  )
 }
